@@ -12,15 +12,15 @@
         }
     ],
     "require": {
-        "php":  ">=5.3.0",
-        "illuminate/database": "*"
+        "php":  ">=5.4.0",
+        "illuminate/database": ">=5.0.0"
     },
     "require-dev": {
         "phpunit/phpunit": "4.2.2"
     },
     "autoload": {
-        "psr-0": {
-            "Corcel": "src/"
+        "psr-4": {
+            "Corcel\\": "src/"
         },
         "classmap": [
             "models"
