@@ -7,7 +7,8 @@ use Illuminate\Database\Eloquent\Model;
 class TermRelationship extends Model
 {
     protected $table = 'term_relationships';
-    protected $primaryKey = array('object_id', 'term_taxonomy_id');
+    protected $primaryKey = ['object_id', 'term_taxonomy_id'];
+    public $timestamps = false;
 
     public function post()
     {
