@@ -2,32 +2,59 @@
 
 namespace Corcel;
 
-use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Model as Eloquent;
 
-class TermTaxonomy extends Model
+class TermTaxonomy extends Eloquent
 {
     protected $table = 'term_taxonomy';
     protected $primaryKey = 'term_taxonomy_id';
     protected $with = array('term');
+    public $timestamps = false;
 
+    /**
+     * Relationship with Term model
+     * @return Illuminate\Database\Eloquent\Relations
+     */
     public function term()
     {
         return $this->belongsTo('Corcel\Term', 'term_id');
     }
 
+    /**
+     * Relationship with parent Term model
+     * @return Illuminate\Database\Eloquent\Relations
+     */
     public function parentTerm()
     {
         return $this->belongsTo('Corcel\TermTaxonomy', 'parent');
     }
 
+    /**
+     * Relationship with Posts model
+     * @return Illuminate\Database\Eloquent\Relations
+     */
     public function posts()
     {
         return $this->belongsToMany('Corcel\Post', 'term_relationships', 'term_taxonomy_id', 'object_id');
     }
 
     /**
-     * Overriding newQuery() to the custom TermTaxonomyBuilder with some interesting methods
+     * Alias from posts, but made quering nav_items cleaner.
+     * Also only possible to use when Menu model is called or taxonomy is 'nav_menu'
      *
+     * @return Illuminate\Database\Eloquent\Relations
+     */
+    public function nav_items()
+    {
+        if ($this->taxonomy == 'nav_menu') {
+            return $this->posts()->orderBy('menu_order');
+        }
+
+        return $this;
+    }
+
+    /**
+     * Overriding newQuery() to the custom TermTaxonomyBuilder with some interesting methods
      * @param bool $excludeDeleted
      * @return Corcel\TermTaxonomyBuilder
      */
@@ -36,15 +63,15 @@ class TermTaxonomy extends Model
         $builder = new TermTaxonomyBuilder($this->newBaseQueryBuilder());
         $builder->setModel($this)->with($this->with);
 
-        if( isset($this->taxonomy) and !empty($this->taxonomy) and !is_null($this->taxonomy) )
+        if (isset($this->taxonomy) and !empty($this->taxonomy) and !is_null($this->taxonomy)) {
             $builder->where('taxonomy', $this->taxonomy);
+        }
 
         return $builder;
     }
 
     /**
      * Magic method to return the meta data like the post original fields
-     *
      * @param string $key
      * @return string
      */
